# Conflicts:
#	src/main/java/com/example/workerrental/repository/SkillRepository.java
#	src/main/java/com/example/workerrental/repository/entity/SkillEntity.java
#	src/main/java/com/example/workerrental/repository/entity/WorkerEntity.java
#	src/main/resources/application.properties

+++ src/main/java/com/example/workerrental/repository/SkillRepository.java
@@ -3,6 +3,5 @@ package com.example.workerrental.repository;
 import com.example.workerrental.repository.entity.SkillEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository<SkillEntity,Long> {
-
+public interface SkillRepository extends JpaRepository<SkillEntity, Integer> {
 }
